Fixed the alignment of the element description
If an element has a long doc string with newlines in it the description
was difficult to read. For example:

  credentialsLife          Credentials lifetime, describing if and how
credentials can be revoked. One of:
* NONE - not applicable
* UNKNOWN - unknown lifetime
* HARDCODED - cannot be invalidated at all, default: Lifetime.NONE

With this commit the description is aligned with the first line of the
description.
  credentialsLife          Credentials lifetime, describing if and how
                           credentials can be revoked. One of:
                           * NONE - not applicable
                           * UNKNOWN - unknown lifetime
                           * HARDCODED - cannot be invalidated at all
                           default: Lifetime.NONE

As a side effect the default and required flags are on a
separated line. I think this makes them easier to spot. But this could
be changed with extra code.