Added Error handling for User errors
Some issues where raised where the users made an error when
executing pytm on the shell.

It seems that the issue is that pytm is not returning a custom error but
just a Python trace.

This commit add the UIError, which will be handled inside the process
function.

UIError is a warpper arround another exception and adds a context
description.

This commit only adds UIError for open calls in pytm, since these are
obvious pitfalls for users.