Added a first draft for #234
In #234 @colesmj suggested to move the import of pydal into the sqlDumb
function.
This commit does this and if the import fails raises an UIError with an
explanation on how to proceed.
The text is just a first draft.

To move the import the function get_table was also moved inside the
sqlDump function.