Added `prerequisites` and `likelihood` to Threat
In threats.json the two properties ("prerequisites", "Likelihood Of Attack") are defined, but are not used in the
rest of pytm.

This commit adds the two properties to the Threat class, so they can be
used by other parts of pytm.

For me this was relevant, since I started to experiment with a different
format for threats mentioned in #237 .
And after exporting threat.json to a markdown format and back into
threat.json these two fields where missing.